refactor: add abstract action trigger
